+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2026-02-06
+
+### Changed
+- `scripts/setup.sh` now downloads binary directly instead of executing remote install script
+
+### Security
+- Removed shell script piping pattern that triggered security scanners
+
 ## [1.4.0] - 2026-02-06
 
 ### Added